CAI2/CAI2/requirements.txt
Flask==2.1.1
Flask-CORS==3.0.10
google-cloud-speech==2.11.0
google-cloud-texttospeech==2.11.0
google-cloud-storage==2.5.0  
